ch2: add error handling to the greeting function
the error is a variable that you create and return with the message.
if there is an error you return in this case an empty message and a new
error with errors.New() function.
if there is no error you return the message and nil, idk what nil is but
it's probably something like no error or null.
in the hello function, if there is an error you should log it, using the
appropriate logger.