ch2: returning greetings for multiple people
To start off, we would be changing the parameter for Hello() function to
accomodate multiple names, which would change the function signature,
and that would break people's code if this was a published module and
people used the Hello() function.

Instead a new function Hellos() is created, the function takes a slice
of names and return a map of names (keys) paired with their greetings
(values).

You specify the type of the key within the bracket and the type of the
value outside the bracket, ex: map[int]string.

We loop through the slice with a for loop, note the slice is also a key,
value data structure, where the index is the key, since the index is not
needed in the loop, we use an underscore _ called blank identifier.

The rest is simple.