ch2: compile and install the app
Running go build in the directory with the go.mod file create a binary
that runs the app.

The app can also be installed globally, by specifying the path where the
binary would be installed, and running go install.

You get the path by running "go list -f '{{.Target}}'".

The path is then exported to the PATH variable, by running
"export PATH=$PATH:/path/to/your/install/directory".

Finally, running the go install command, a new file is created in that
directory, and running hello in this ex in the terminal will run it.