ch3: multi-module workspaces
Basically, we create a workspace in which multiple modules live.

First we had the hello module that used a remote function from golang to
reverse strings, it had to go download the whole golang.org/x/example
and put it in a cache.

We tell the workspace to use this module as a main module with
go work init ./hello, now within our workspace we can run use ./hello to
run the hello module, even if we are not inside the directory.

Basically it initializes a Go workspace with the given module as the
first module in that workspace.

Next we clone the example repo and put it in the workspace as well.

We add this module to the workspace using the go work use
./example/hello.

Basically, this command is used to add an existing module to the
workspace that has already been initialized with go work init.

Then we modified the reverse file in the example/hello directory, and
added a reverse integer function.

Because we added example/hello to the workspace, the workspace will
establish a link between golang.org/x/example/hello and the local
example/hello directory, this ensures Go resolves the
golang.org/x/example/hello import path to the local example/hello
directory.