ch5: Developing a RESTful API with Go and Gin
Import the needed modules, mainly http and gin.

Run go get . to get the gin library from github.

Create a struct for the album, and a slice with some albums.

The main function declares a gin server, and sets the routes and handler
functions for each route, there are 3 routes 2 GETS for all albums and a
single album, and a POST to send one album.

The getAlbums function takes a param c of type gin.Context which is
basically the http request I think, and in the function
c.IndentedJSON method is called which basically sends a response with an
http status of 200 and a json body of the albums slice.

The getAlbumByID also take gin context param, retrieves the id from the
parameters of the request with c.Params("id") and assigns it to an id
variable, and then loops through the albums slice with a conditional to
match the id value with the album.id value, if found its returned like
in getAlbum with a c.IndentedJSON(http.StatusOK,a) a being the album in
the response body.

If no match has been found a status not found is returned with a album
not found message.

Finally, postAlbums also take c *gin.Context as a param, creates a
newAlbum struct of type album, and binds the content of the body of the
request to newAlbum, if there is no error it just returns, idk what
happens when there is an error doesnt look like we send smth back.

After a 'guaranted' success, we append the newAlbum to albums, and send
with the response a statusCreated, and newAlbum in the body.